refactor(InputView, OutputView): Move InputView&OutputView to view
